api/requirements.txt
Django==4.2.1
django-extensions==3.2.1
djangorestframework==3.14.0
djoser==2.2.0
Pillow==9.5.0
